Support the dedicated server console on Linux using SDL
Requires SDL2_ttf. It compiles, but I can't confirm it works until the
engine initialization process works on Linux. The original Windows
console is still there. In the future it could be deprecated in favor of
the SDL version.

TODO: Find font dynamically using fontconfig. Right now it looks for
/usr/share/fonts/TTF/arial.ttf